warnings

+++ 7zpp/FileSys.cpp
@@ -189,7 +189,7 @@ CComPtr< IStream > FileSys::OpenFileToRead( const TString& filePath )
 	const WCHAR* filePathStr = filePath.c_str();
 #else
 	WCHAR filePathStr[MAX_PATH];
-	MultiByteToWideChar( CP_UTF8, 0, filePath.c_str(), filePath.length() + 1, filePathStr, MAX_PATH );
+	MultiByteToWideChar( CP_UTF8, 0, filePath.c_str(), (int)filePath.length() + 1, filePathStr, MAX_PATH );
 #endif
 	if ( FAILED( SHCreateStreamOnFileEx( filePathStr, STGM_READ, FILE_ATTRIBUTE_NORMAL, FALSE, NULL, &fileStream ) ) )
 	{
@@ -207,7 +207,7 @@ CComPtr< IStream > FileSys::OpenFileToWrite( const TString& filePath )
 	const WCHAR* filePathStr = filePath.c_str();
 #else
 	WCHAR filePathStr[MAX_PATH];
-	MultiByteToWideChar( CP_UTF8, 0, filePath.c_str(), filePath.length() + 1, filePathStr, MAX_PATH );
+	MultiByteToWideChar( CP_UTF8, 0, filePath.c_str(), (int)filePath.length() + 1, filePathStr, MAX_PATH );
 #endif
 	if (FAILED(SHCreateStreamOnFileEx(filePathStr, STGM_CREATE | STGM_WRITE, FILE_ATTRIBUTE_NORMAL, TRUE, NULL, &fileStream)))
 	{
